Moved cloud_api.py to the root directory
